Fix off-by-one in height rendering
The board is stored with low indices representing the bottom,
so the y-index is subtracted from the height to position low
cells at the bottom. Because the board is 0-indexed, too little was
subtracted, causing rows to render too low, e.g. - the 0th row drew
cells with their tops at canvas.height - 0